refactor model creation to files
